plugged constant memory leak. altered temporary_eval to avoid memory leak
