+++ tenncor/src/graph/leaf/placeholder.ipp
@@ -91,7 +91,7 @@ placeholder<T>& placeholder<T>::operator = (tensor<T>& data)
 template <typename T>
 inode<T>* placeholder<T>::get_leaf (variable<T>*)
 {
-	return get_shared_zero<T>();
+	return constant<T>::get_shared_zero();
 }
 
 template <typename T>
