+++ tenncor/tests/include/mocks/mock_connector.h
@@ -73,7 +73,7 @@ protected:
 		return new mock_connector(std::move(*this));
 	}
 
-	virtual void summarize (std::vector<typename iconnector<double>::conn_summary>&) const {}
+	virtual std::vector<typename iconnector<double>::conn_summary> summarize (void) const { return {}; }
 };
